Cleaning up, create ExportLogsViewModel, extract string resources
