+++ DittoExportLogs/src/main/java/live/ditto/dittoexportlogs/ExportLogs.kt
@@ -2,8 +2,6 @@ package live.ditto.dittoexportlogs
 
 import android.os.Build
 import android.os.Environment
-import android.util.Log
-import android.widget.Toast
 import androidx.annotation.RequiresApi
 import androidx.compose.foundation.layout.Row
 import androidx.compose.foundation.layout.Spacer
@@ -19,21 +17,20 @@ import androidx.compose.runtime.getValue
 import androidx.compose.runtime.mutableStateOf
 import androidx.compose.runtime.remember
 import androidx.compose.runtime.rememberCoroutineScope
-import androidx.compose.runtime.setValue
 import androidx.compose.ui.Modifier
 import androidx.compose.ui.platform.LocalContext
 import androidx.compose.ui.res.stringResource
 import androidx.compose.ui.unit.dp
 import kotlinx.coroutines.launch
-import live.ditto.DittoError
-import live.ditto.DittoLogger
 import java.io.File
 import java.time.LocalDateTime
 import java.time.format.DateTimeFormatter
 
 @RequiresApi(Build.VERSION_CODES.O)
 @Composable
-fun ExportLogs(onDismiss: () -> Unit) {
+fun ExportLogs(onDismiss: () -> Unit, viewModel: ExportLogsViewModel = ExportLogsViewModel()) {
+
+    val isLoading by remember { mutableStateOf(viewModel.isLoading)}
 
     val title = stringResource(R.string.export_logs)
 
@@ -41,7 +38,6 @@ fun ExportLogs(onDismiss: () -> Unit) {
     val cancelText = stringResource(R.string.cancel)
     val modifier = Modifier
     val coroutineScope = rememberCoroutineScope()
-    var isLoading by remember { mutableStateOf(false) }
     val context = LocalContext.current
 
     val now = LocalDateTime.now().format(DateTimeFormatter.ofPattern("yyyyMMdd_HHmmss"))
@@ -58,7 +54,7 @@ fun ExportLogs(onDismiss: () -> Unit) {
             Text(title)
         },
         text = {
-            Text("Logs will be exported to: \nDownloads/$fileName\"")
+            Text(stringResource(R.string.export_logs_message, fileName))
         },
 
         confirmButton = {
@@ -67,22 +63,7 @@ fun ExportLogs(onDismiss: () -> Unit) {
 
 
                     coroutineScope.launch {
-                        isLoading = true
-                        try {
-                            val sizeWritten = DittoLogger.exportToFile(filePath)
-
-                            val sizeWrittenInKB = sizeWritten / 1024u
-                            val toastText = "Exported file size: $sizeWrittenInKB kB"
-                            Toast.makeText(context, toastText, Toast.LENGTH_LONG).show()
-                        }
-                        catch (e: DittoError.IoError) {
-                            Log.e("DittoTools LogExporter", e.reason.toString())
-                            Toast.makeText(context, "Error exporting logs: ${e.message}", Toast.LENGTH_LONG).show()
-                        }
-                        finally {
-                            isLoading = false
-                            onDismiss()
-                        }
+                       viewModel.exportLogs(filePath, context, onDismiss)
                     }
                 }
             ) {
